updated code to run on kepler

+++ cuda/benchmarks/vector_sum.cuh
@@ -42,11 +42,11 @@ class VectorSum : public Benchmark {
     std::string print_result(bool short_form = false);
 
    private:
-    double *x, *res_tmp;
-    double res = 0.0;
-    double *x_d, *res_tmp_d;
+    float *x, *res_tmp;
+    float res = 0.0;
+    float *x_d, *res_tmp_d;
     int B = 0;
-    double cpu_result;
+    float cpu_result;
 
     // Implementations of the algorithm;
     void vector_sum_0(int iter);
